Added option to the option "calculate energy cost"
The option allows users to choose if they want to calculate the cost if you reuse an ingredient made with a product. 
Example, crushing iron ore creates crushed iron and crushed stone. To make the iron ore you need 20 crushed stone. This option allows you to calculate as if the recipe made crushed iron but took 19 crushed stone.